Separating switch transformer from transformer
